requirements.txt
python-dotenv
fastapi==0.122.0
uvicorn[standard]==0.38.0
jinja2==3.1.6
python-multipart==0.0.20
groq==0.36.0